Inline config loading
