+++ src/useConfig.ts
@@ -1,7 +1,3 @@
-import { useEffect, useMemo } from "react";
-import useSWRImmutable from "swr/immutable";
-import { jsonFetcherWithErrorHandling } from "./fetcher";
-
 /**
  * Defines a set of metadata for a certain chain.
  *
@@ -197,49 +193,3 @@ export type OtterscanConfig = {
  * Default location for fetching the config file.
  */
 export const DEFAULT_CONFIG_FILE = "/config.json";
-
-/**
- * Fetches the config file and parse it into proper config object.
- *
- * On development environment, allows the config to be set by overriding
- * specific params over the default "/public/config.json" by using Vite's
- * env mechanism: https://vitejs.dev/guide/env-and-mode.html
- */
-export const useConfig = (
-  configURL: string = DEFAULT_CONFIG_FILE,
-): OtterscanConfig | undefined => {
-  const { data } = useSWRImmutable(configURL, jsonFetcherWithErrorHandling);
-  const config = useMemo(() => {
-    if (data === undefined) {
-      return undefined;
-    }
-
-    // Override config for local dev
-    const _config: OtterscanConfig = { ...data };
-    if (import.meta.env.DEV) {
-      _config.erigonURL = import.meta.env.VITE_ERIGON_URL ?? _config.erigonURL;
-      _config.beaconAPI =
-        import.meta.env.VITE_BEACON_API_URL ?? _config.beaconAPI;
-      _config.assetsURLPrefix =
-        import.meta.env.VITE_ASSETS_URL ?? _config.assetsURLPrefix;
-      _config.experimental =
-        import.meta.env.VITE_EXPERIMENTAL ?? _config.experimental;
-      if (import.meta.env.VITE_EXPERIMENTAL_FIXED_CHAIN_ID !== undefined) {
-        _config.experimentalFixedChainId = parseInt(
-          import.meta.env.VITE_EXPERIMENTAL_FIXED_CHAIN_ID,
-        );
-      }
-    }
-    return _config;
-  }, [data]);
-
-  useEffect(() => {
-    if (data === undefined) {
-      return;
-    }
-    console.info("Loaded app config");
-    console.info(config);
-  }, [config]);
-
-  return config;
-};
